Redesign edit mode.
Rather than try to combine run mode and edit mode, this commit introduces
a button that enters and exits edit mode.

We also put the edit field (which adds new rows) at the bottom of the
list and adding new rows appends them to the end instead of the top.