Fixed issue with Chrome.
GraphQL subscriptions weren't working on Chrome browsers. Beau and I
found out our GraphQL library didn't specify the `sec-websocket-protocol`
header when upgrading to a websocket and our service was always returning
it. Chrome doesn't like that.

Beaus found a way to add this header to our request, which is the content
of this commit.