ADD - implement `graphql_dpm_service.monitorSettingProperty()`
This function simply calls `graphql_dpm_service.monitorDevices()`. It
might seem that this function should append ".SETTING" to the devices
in the list, but we need to support EPICS devices, too, so that won't
work. The parameter page may need more intelligence to decide how to
specify a device's setting.